soc/mediatek/mt8192: add rtc MT6359P driver
Add rtc MT6359P driver for rtc init and rtc eosc calibration. Refactor
mt8173 and mt8183 code by extracting common API. Move rtc_read and
rtc_write to each SoC folder, because mt8173 and mt8183 access rtc via
pmic wrapper, while mt8192 accesses it via pmif.

Reference datasheet:
  Document No: RH-D-2018-0101.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/46395
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/mediatek/mt8173/rtc.c
@@ -1,8 +1,8 @@
 /* SPDX-License-Identifier: GPL-2.0-only */
 
 #include <delay.h>
-#include <soc/rtc_common.h>
 #include <soc/rtc.h>
+#include <soc/rtc_common.h>
 #include <soc/mt6391.h>
 #include <soc/pmic_wrap.h>
 #include <types.h>
@@ -59,7 +59,7 @@ static int rtc_lpd_init(void)
 }
 
 /* rtc init check */
-int rtc_init(u8 recover)
+int rtc_init(int recover)
 {
 	int ret;
 
@@ -84,6 +84,7 @@ int rtc_init(u8 recover)
 		goto err;
 	}
 
+	/* In recovery mode, we need delay for register setting. */
 	if (recover)
 		mdelay(1000);
